requirements.txt
pandas>=1.5.3
numpy>=1.24.4
pyyaml>=6.0
openpyxl>=3.1.2
matplotlib>=3.7.1
streamlit>=1.45.1
plotly>=5.9.0
